# docs/core/CORE_API_OPENAPI.yaml
openapi: 3.1.1
info:
  title: Вузы России — API ядра-бекенда
  version: 0.1.0
  description: 'Предложенный контракт MVP от 2026-09-24. Не описание текущей реализации. Только операции ядра, включая
    API, вызываемое backend сервисов. Доменные API других компонентов не входят. CORE_API_SPEC.md дополняет HTTP-контракт
    правилами JWT, изоляции, транзакций и bootstrap. Private операции имеют отдельный server и AND-требование machine
    + actor; на публичном ingress всегда 404. x-required-permission, x-required-machine-scope и x-network-access
    — документационные extensions, не автоматическое обеспечение доступа. Форматы JSON: snake_case, UUID string,
    UTC. Формат документа — OpenAPI 3.1.1 (https://spec.openapis.org/oas/v3.1.1.html).'
servers:
- url: https://core.example
  description: Публичный HTTPS origin ядра; заменить при развёртывании.
security:
- CoreBearer: []
...
paths:
...
  /api/v1/institution/{institution_id}/internal:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Private administration
      summary: Прочитать настройки вуза
      operationId: getInstitutionSettings
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInstitution'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Требуемый permission actor: institution.read.'
      x-required-permission: institution.read
    patch:
      tags:
      - Private administration
      summary: Изменить имя или язык вуза
      operationId: updateInstitutionSettings
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInstitution'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Регистрация/одобрение/приостановка вуза —
        вне HTTP API MVP. Требуемый permission actor: institution.update.'
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionPatch'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      x-required-permission: institution.update
  /api/v1/institution/{institution_id}/internal/service-types:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Private administration
      summary: Получить разрешённые типы сервисов
      operationId: listServiceTypes
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTypeList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Четыре фиксированных типа MVP. Каталог управляется
        платформой; создания произвольного типа нет. Требуемый permission actor: institution.read.'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      x-required-permission: institution.read
  /api/v1/institution/{institution_id}/internal/members:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    get:
      tags:
      - Private members
      summary: Прочитать список участников вуза
      operationId: listInstitutionMembers
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Требуемый permission actor: members.read.'
      x-required-permission: members.read
    post:
      tags:
      - Private members
      summary: Добавить зарегистрированного пользователя и профили
      operationId: addInstitutionMember
      responses:
        '201':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
            Location:
              $ref: '#/components/headers/Location'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Вход через MAX должен быть выполнен заранее.
        Повтор пары вуз+пользователь — 409. Добавление admin само по себе не выдаёт административных ролей. Требуемый
        permission actor: members.manage.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreate'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      x-required-permission: members.manage
  /api/v1/institution/{institution_id}/internal/members/{user_id}:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/user_id'
    get:
      tags:
      - Private members
      summary: Прочитать членство и его ETag
      operationId: getInstitutionMember
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Требуемый permission actor: members.read.'
      x-required-permission: members.read
    delete:
      tags:
      - Private members
      summary: Удалить членство, его профили, назначения и сессии
      operationId: removeInstitutionMember
      responses:
        '204':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Аккаунт и членство в других вузах сохраняются.
        Данные образовательных сервисов автоматически не удаляются. Последний owner — 409. Требуемый permission
        actor: members.manage.'
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      x-required-permission: members.manage
  /api/v1/institution/{institution_id}/internal/members/{user_id}/profiles:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/user_id'
    put:
      tags:
      - Private members
      summary: Заменить набор профилей участника
      operationId: replaceMemberProfiles
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '428':
          $ref: '#/components/responses/Error428'
        '500':
          $ref: '#/components/responses/Error500'
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. If-Match берётся из GET membership. Набор
        непустой. Снятые профили теряют роли/сессии во всех экземплярах вуза. Последний owner защищён. Требуемый
        permission actor: members.manage.'
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilesInput'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      x-required-permission: members.manage
...
  /api/v1/institution/{institution_id}/internal/services/{service_id}/roles:
    parameters:
    - $ref: '#/components/parameters/institution_id'
    - $ref: '#/components/parameters/service_id'
    get:
      tags:
      - Private roles
      summary: Прочитать роли выбранного экземпляра вуза
      operationId: listInstitutionServiceRoles
      responses:
        '200':
          description: Успешно.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      servers:
      - url: https://core.private.example
        description: Приватный listener; доступ только workload администрирования.
      security:
      - MachineBearer: []
        ActorToken: []
      x-network-access: private-administration-only
      description: 'Public ingress всегда отвечает 404 до аутентификации. На private listener обязательны оба токена,
        совпадение вуза/экземпляра администрации и свежие права actor. Требуемый permission actor: services.read.'
      x-required-permission: services.read
...
components:
  securitySchemes:
    CoreBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Пользовательский token_use=core_access, aud=core-api; живая sid обязательна.
    MachineBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token_use=machine_access, aud=core-internal. Credential и tenant проверяются на каждом запросе.
        Не пользовательский токен.
    ServiceBasic:
      type: http
      scheme: basic
      description: 'Только HTTPS token exchange: client_id как username, client_secret как password.'
    ActorToken:
      type: apiKey
      in: header
      name: X-Actor-Token
      description: 'JWT без префикса Bearer: service_access человека для экземпляра administration, профиль admin.
        AND с MachineBearer; актуальный RBAC из БД.'
  parameters:
...
    institution_id:
      name: institution_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  headers:
    RequestId:
      description: UUID запроса, назначенный ядром.
      schema:
        $ref: '#/components/schemas/UUID'
    NoStore:
      description: Ответ нельзя сохранять в HTTP-кэше.
      schema:
        type: string
        const: no-store
    ETag:
      description: Strong ETag ресурса для If-Match.
      schema:
        type: string
        examples:
        - '"rev-7"'
...
  responses:
    Error400:
      description: Невалидный JSON, cursor или синтаксис запроса.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    AdminInstitution:
      type: object
      additionalProperties: false
      required:
      - id
      - titles
      - default_locale
      - status
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        titles:
          $ref: '#/components/schemas/LocalizedText'
        default_locale:
          $ref: '#/components/schemas/Locale'
        status:
          $ref: '#/components/schemas/InstitutionStatus'
...